refactor: migrate categories feature to feature slice
- Create src/features/categories/ with data.ts, index.ts, and tests
- Move STANDARD_CATEGORIES and getCategoryById to feature slice
- Update imports across codebase to use @/features/categories
- Create backward-compatible re-export in src/data/standardCategories.ts
- All 890 tests pass
