+++ src/components/settings/ShoppingListExport.tsx
@@ -2,7 +2,7 @@ import { useTranslation } from 'react-i18next';
 import { useInventory } from '@/shared/hooks/useInventory';
 import { Button } from '@/shared/components/Button';
 import type { InventoryItem } from '@/shared/types';
-import { STANDARD_CATEGORIES } from '@/data/standardCategories';
+import { STANDARD_CATEGORIES } from '@/features/categories';
 import styles from './ShoppingListExport.module.css';
 
 export function ShoppingListExport() {
