+++ src/data/standardCategories.test.ts
@@ -1,47 +1,18 @@
-import { STANDARD_CATEGORIES, getCategoryById } from './standardCategories';
+/**
+ * @deprecated Tests moved to '@/features/categories/data.test.ts'
+ * This file is kept for backward compatibility
+ */
+import { STANDARD_CATEGORIES, getCategoryById } from '@/features/categories';
 
-describe('standardCategories', () => {
-  describe('STANDARD_CATEGORIES', () => {
-    it('should contain 9 standard categories', () => {
-      expect(STANDARD_CATEGORIES).toHaveLength(9);
-    });
-
-    it('should have water-beverages as first category', () => {
-      expect(STANDARD_CATEGORIES[0].id).toBe('water-beverages');
-      expect(STANDARD_CATEGORIES[0].icon).toBe('💧');
-    });
-
-    it('should have cash-documents as last category', () => {
-      expect(STANDARD_CATEGORIES[8].id).toBe('cash-documents');
-      expect(STANDARD_CATEGORIES[8].icon).toBe('💰');
-    });
-
-    it('should have all categories marked as not custom', () => {
-      STANDARD_CATEGORIES.forEach((category) => {
-        expect(category.isCustom).toBe(false);
-      });
-    });
+describe('standardCategories (backward compatibility)', () => {
+  it('should re-export STANDARD_CATEGORIES from features/categories', () => {
+    expect(STANDARD_CATEGORIES).toHaveLength(9);
+    expect(STANDARD_CATEGORIES[0].id).toBe('water-beverages');
   });
 
-  describe('getCategoryById', () => {
-    it('should return category when id exists', () => {
-      const category = getCategoryById('water-beverages');
-      expect(category).toBeDefined();
-      expect(category?.id).toBe('water-beverages');
-      expect(category?.name).toBe('Water & Beverages');
-    });
-
-    it('should return food category', () => {
-      const category = getCategoryById('food');
-      expect(category).toBeDefined();
-      expect(category?.id).toBe('food');
-      expect(category?.icon).toBe('🍽️');
-    });
-
-    it('should return undefined for non-existent category', () => {
-      // @ts-expect-error Testing invalid input
-      const category = getCategoryById('non-existent');
-      expect(category).toBeUndefined();
-    });
+  it('should re-export getCategoryById from features/categories', () => {
+    const category = getCategoryById('water-beverages');
+    expect(category).toBeDefined();
+    expect(category?.id).toBe('water-beverages');
   });
 });
