+++ src/features/categories/index.ts
@@ -1,0 +1,2 @@
+// Data
+export { STANDARD_CATEGORIES, getCategoryById } from './data';
